code/requirements.txt
absl-py==1.4.0
alabaster==0.7.13
albumentations==1.2.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.2.0
audioread==3.0.0
autograd==1.5
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
biopython==1.81
bleach==6.0.0
blis==0.7.9
bokeh==2.4.3
branca==0.6.0
bs4==0.0.1
CacheControl==0.12.11
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.0
cloudpickle==2.2.1
cmake==3.25.2
cmdstanpy==1.1.0
colorcet==3.0.1
colorlover==0.3.0
commonmark==0.9.1
community==1.0.0b1
confection==0.0.4
cons==0.4.5
contextlib2==0.6.0.post1
convertdate==2.4.0
cufflinks==0.17.3
cvxopt==1.3.0
cvxpy==1.3.1
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
datascience==0.17.6
db-dtypes==1.0.5
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
dlib==19.24.0
dm-tree==0.1.8
docker==6.0.1
docutils==0.16
dopamine-rl==1.0.5
earthengine-api==0.1.345
easydict==1.10
ecos==2.0.12
editdistance==0.5.3
entrypoints==0.4
ephem==4.1.4
epitopepredict==0.5.0
et-xmlfile==1.1.0
etuples==0.3.8
fastai==2.7.11
fastcore==1.5.28
fastdownload==0.0.7
fastjsonschema==2.16.3
fastprogress==1.0.3
fastrlock==0.8.1
filelock==3.10.0
firebase-admin==5.3.0
fix-yahoo-finance==0.0.22
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
folium==0.14.0
fonttools==4.38.0
future==0.18.3
gast==0.2.2
gdown==4.6.4
geographiclib==2.0
geopy==2.3.0
gin-config==0.5.0
glob2==0.7
google==2.0.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.4.2
google-cloud-bigquery-storage==2.19.0
google-cloud-core==2.3.2
google-cloud-datastore==2.11.1
google-cloud-firestore==2.7.3
google-cloud-language==2.6.1
google-cloud-storage==2.7.0
google-cloud-translate==3.8.4
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
googledrivedownloader==0.4
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.48.2
gspread==3.4.2
gspread-dataframe==3.0.8
gunicorn==20.1.0
gym==0.25.2
gym-notices==0.0.8
h5netcdf==1.1.0
h5py==3.8.0
HeapDict==1.0.1
hijri-converter==2.2.4
holidays==0.21.13
holoviews==1.15.4
html5lib==1.1
htmlmin==0.1.12
httpimport==1.3.0
httplib2==0.21.0
humanize==4.6.0
hyperopt==0.2.7
idna==3.4
ifeature==0.0.6
ImageHash==4.3.1
imageio==2.25.1
imagesize==1.4.1
imbalanced-learn==0.10.1
imblearn==0.0
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
imutils==0.5.4
inflect==6.0.2
intel-openmp==2023.0.0
ipykernel==5.3.4
ipython==7.9.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.1.1
jsonschema==4.3.3
jupyter-client==6.1.12
jupyter-console==6.1.0
jupyter-server==1.23.6
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.13
keras==2.11.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-vis==0.4.1
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
langcodes==3.3.0
lazy_loader==0.1
libclang==15.0.6.1
librosa==0.10.0.post2
lightgbm==3.3.5
llvmlite==0.39.1
locket==1.0.0
logical-unification==0.4.5
LunarCalendar==0.0.9
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mhcflurry==2.0.6
mhcgnomes==1.8.4
miniKanren==1.0.3
missingno==0.5.2
mistune==0.8.4
mkl==2019.0
mlxtend==0.14.0
more-itertools==9.1.0
moviepy==0.2.3.5
mpmath==1.3.0
msgpack==1.0.5
multimethod==1.9.1
multipledispatch==0.6.0
multitasking==0.0.11
murmurhash==1.0.9
music21==5.5.0
mypy-extensions==1.0.0
natsort==5.5.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==6.5.4
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
nibabel==3.0.2
nltk==3.8.1
notebook==6.3.0
notebook_shim==0.2.2
np-utils==0.6.0
numba==0.56.4
numexpr==2.8.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.0.10
opt-einsum==3.3.0
osqp==0.6.2.post0
packaging==23.0
palettable==3.3.0
pandas==1.3.5
pandas-datareader==0.10.0
pandas-gbq==0.17.9
pandas-profiling==3.2.0
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
partd==1.3.0
pathlib==1.0.1
pathy==0.10.1
patsy==0.5.3
pep517==0.13.0
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==8.4.0
pip-tools==6.6.2
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
plotly==5.13.1
pluggy==0.7.1
pooch==1.6.0
portpicker==1.3.9
prefetch-generator==1.0.3
preshed==3.0.8
prettytable==3.6.0
progressbar2==3.38.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==2.0.10
prophet==1.1.2
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
py4j==0.10.9.7
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybiolib==1.1.809
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.17
pyct==0.5.0
pydantic==1.10.6
pydata-google-auth==1.7.0
pydot==1.4.2
pydot-ng==2.0.0
pydotplus==2.0.2
PyDrive==1.3.1
pyerfa==2.0.0.2
Pygments==2.6.1
PyJWT==2.6.0
PyMeeus==0.5.12
pymystem3==0.2.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==3.6.4
python-dateutil==2.8.2
python-louvain==0.16
python-slugify==8.0.1
python-utils==3.5.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyWavelets==1.3.0
PyYAML==5.4
pyzmq==23.2.1
qdldl==0.1.5.post3
qudida==0.0.4
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==12.6.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
scs==3.2.2
seaborn==0.12.2
Send2Trash==1.8.0
serializable==0.2.1
shapely==2.0.1
simplejson==3.18.4
six==1.16.0
sklearn==0.0.post1
sklearn-pandas==2.2.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.47
sqlparse==0.4.3
srsly==2.4.6
statsmodels==0.13.5
tables==3.7.0
tabulate==0.8.10
tangled-up-in-unicode==0.2.0
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-gcs-config==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.12.0
tensorflow-probability==0.19.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tmhmm.py @ file:///home/ubuntu/Desktop/nerve/tmhmm.py
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchsummary==1.5.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
tweepy==4.13.0
typechecks==0.1.0
typer==0.7.0
typing-inspect==0.5.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
vega-datasets==0.9.0
visions==0.7.4
wasabi==1.1.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==3.6.2
wordcloud==1.8.2.2
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yellowbrick==1.5
zict==2.2.0
zipp==3.15.0
